[GetDictionaries] - part 3
